requirements.txt
fastmcp>=2.0.0
pydantic>=2.7.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
email-validator>=2.1.0
serpapi>=0.1.0
google-search-results>=2.4.0